requirements.txt
Django==3.2.12
environs==9.5.0
django-registration==3.3
python-dateutil==2.8.2
Pillow==9.4.0